ci: allow dependabot PRs to update security-events
Without this permission PRs triggered by dependabot or other users won't
be able to run the `github/codeql-action/upload-sarif` action which
gives the following error:

"This run of the CodeQL Action does not have permission to access Code
Scanning API endpoints. As a result, it will not be opted into any
experimental features. This could be because the Action is running on a
pull request from a fork. If not, please ensure the Action has the
'security-events: write' permission.
Details: Resource not accessible by integration"